Use posix threads inside in order to support a faster network. Fix fake upload problem(with a readback added)

+++ src/http.h
@@ -1,10 +1,10 @@
 #ifndef HTTPH
 #define HTTPH
-
+#include <sys/types.h>
 int httpPut(char* pAddress, int pPort, char* pRequest, unsigned long contentSize);
 int httpGet(char* pAddress, int pPort, char* pRequest);
 int httpRecv(int pSockId, char* pOut, int pOutSize);
-int httpSend(int pSockId, char* pOut, int pOutSize);
+ssize_t httpSend(int pSockId, char* pOut, int pOutSize);
 int httpLastError();
 void httpClose(int pSockId);
 int httpGetRequestSocket(const char *urlToDownload);
